pyproject.toml
[tool.poetry]
name = "FraudGuard"
version = "0.0.1"
description = "Fraud detection project to classify and monitor suspicious transactions."
authors = ["JavithNaseem-J <javithnaseem.j@gmail.com>"]
packages = [{ include = "FraudGuard", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"

# Core dependencies
numpy = "^1.24.0"
pandas = "^1.5.0"
scipy = "^1.10.0"
joblib = "^1.3.0"
pydantic = "^2.5.0"

# API and web framework
fastapi = "^0.103.0"
uvicorn = { version = "^0.23.0", extras = ["standard"] }
python-multipart = "^0.0.6"
jinja2 = "^3.1.0"

# Configuration and utilities
python-box = "^7.0.0"
pyyaml = "^6.0"
ensure = "^1.0.0"
tqdm = "^4.65.0"
gdown = "^4.7.0"

# ML libraries
scikit-learn = "^1.3.0"
imbalanced-learn = "^0.11.0"
lightgbm = "^3.3.0"
catboost = "^1.2.0"
xgboost = "^1.7.0"
optuna = "^3.3.0"
ipywidgets = "^8.0.0"

# Cloud and data management
boto3 = "^1.28.0"
dvc = "^3.15.0"
dvc-s3 = "^2.23.0"

# MLOps and experiment tracking
mlflow = "^2.5.0"
dagshub = "^0.3.0"

# Visualization libraries
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
shap = "^0.44.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
httpx = "^0.24.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"